Upgrade phan and unbreak master
Follow-up: I606d6205ee22e48ca6df04c14954ed7cefcadd64 which left an issue unfixed

Also fixes a major bug in SpecialImageRating which caused the 'categorylinks'
table to never be used in SELECTs.
And also a bug where a locale-formatted number was stored in cache and passed
around, and notably it was fed to VoteStars::displayStars, which expects an
integer and will probably fail unexpectedly if given a localised number.

Change-Id: I5939d05e04527dfb20ce5eafa3a6657cdb0f4def