Fix annoying rounding bug with the vote average on Special:ImageRating
Previously, inconsistent with VoteNY (which this extension uses and depends on (!)), we'd display a "raw"er count like 3.3333333333333 instead of a formatted, rounded count like 3.33.
VoteNY's Vote#getAverageVote ensures that only two decimals are shown...but ImageRating does *not* call it since it gets the raw count directly from the DB. (TODO: should it? Would that be more efficient in terms of performance since the VoteNY method has caching?)
Thus we need to do that formatting ourselves before passing on the number to MW's i18n formatting functions.

Change-Id: I028d71f73368b66075d847a595143a4fafc7716d